Added 2D matrix behaviour
When batch processing multiple files at once using the -o flag, the program will output the images into a single 2D matrix which allows the images to be used to create animations or video on the GBA. E.g. './png2gba -p -o out.h test{1..2}.png' will allow the main.c file to loop through all entries in the 2D matrix creating an animation in the process.

Default behaviour for the palette flag -p has also been changed, the palette output is now printed into a separate file and given an include to the associated file. This makes the files, especially in the case of large 2D matrices easier to browse and easier to visualize.

Commit also includes extra information and details into the header files such as include guards, number of entries in the matrix and includes for associated files so the compiler will warn if a file is missing during compilation.