Removing ctype library
Included <ctype.h> library to calculate the maximum size for variables in order to be able to locate them on the stack instead of on the heap. For now, I have implemented practically all variables that take changing/user dependent input as heap based variables using pointers instead, thereby circumventing this problem.